backend/requirements.txt
fastapi
uvicorn
google-cloud-bigquery
pandas
pydantic-settings
scikit-learn
pytest
httpx
